requirements.txt
streamlit>=1.38.0
anthropic>=0.40.0
PyPDF2>=3.0.0
Pillow>=10.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
